home/serializers.py

from rest_framework import serializers

from home.models import *


class ProjectSerializer(serializers.ModelSerializer):
    class Meta:
        model = Project
        fields = '__all__'
        



