requirements.txt
asgiref==3.4.1
Django==3.2.5
django-debug-toolbar==2.2
django-phone-field==1.8.1
djangorestframework==3.12.4
gunicorn==20.0.4
Pillow==8.3.1
psycopg2-binary==2.8.3
pytz==2019.3
sqlparse==0.3.0
typing-extensions==3.10.0.0
